Large changeset: c->cpp, function trimming->selective function adding
Change the way that hardware design files are built. Rather than taking
the entire AST of a source package file and trimming unreachable
functions, we build a template file and then codegen into it the
reachable functions. This is far better because the declarations in
pycparser's dummy header files were messing up HLS.

To support this we change hardware to use C rather than C++ because the
generated template must now include Main__.h which declares C binding
for all functions and messed up with a .cpp top level. C is actually
preferrable for hardware, but we need to test to ensure that the
synthesised hardware is actually working as previous.