Modified kernel driver to use two BARs.
This is to work around an Intel board refusing to boot with a 1GB BAR.

BAR0 is now sized the same as the on-board RAM for the FPGA, and exposes
only the FPGA's DDR.

BAR1 is a 128MB BAR (to encompass 0x20000000 - 0x27FFFFFF) for the peripherals.
This BAR can easily be reduced in size later with some address remapping.