Added fix for buggy HLS clock statements. The fix removes create_clock statements from synthesized checkpoints.
The fix is because synthesized HLS components include constraints which define a clock on the ap_clk pin
using the create_clock XDC statement. This causes problems when connected to the PCI Express clock (which also
has the create_clock statement applied to it automatically) because the tools now assume that a new clock is created at ap_clk,
but also connected to the PCI-E clock. This then causes the tools to assume a clock domain crossing occurs at ap_clk, causing
timing problems. Moreover, this also appears to cause the tools to analyse the delay that the clock experiences when transiting
from the PCI-E block to the appropriate BUFG (i.e. the GTX -> BUFG path), but assumes that ap_clk does not have any "pre" delay
like this, hence also causing the tools to assume a large amount of clock skew.

There does not seem to be a way to disable the generation of this constraint within HLS (removing create_clock from the build TCL
simply causes it to assume a 10ns clock), or to remove constraints within Vivado's TCL interface, hence the cleanest solution
is to unzip the checkpoint and manually remove it.